requirements.txt
asgiref==3.2.3
Django==2.2.5
django-braces==1.13.0
django-model-utils==4.0.0
django-online-users==0.3
django-private-chat==0.2.2
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
websockets==3.2
